% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-calcLeafAreaIndex.R
\name{calcLeafAreaIndex}
\alias{calcLeafAreaIndex}
\title{Calculate Leaf Area Index (LAI)}
\usage{
calcLeafAreaIndex(LAI, LAIp, LAIp_previousDay, stress, stress_nocivity)
}
\arguments{
\item{LAI}{\link{numeric} The current Leaf Area Index.}

\item{LAIp}{\link{numeric} The potential Leaf Area Index growth for the current
day.}

\item{LAIp_previousDay}{\link{numeric} The potential Leaf Area Index growth for
the previous day.}

\item{stress}{\link{numeric} A stress factor affecting the growth of LAI (e.g.,
water stress, nutrient stress).}

\item{stress_nocivity}{\link{numeric} A factor representing the nocivity
(severity) of the stress.}
}
\value{
\link{numeric} The updated Leaf Area Index after accounting for potential
growth and stress factors.
}
\description{
This function calculates the Leaf Area Index (LAI) based on the current LAI,
potential LAI growth, previous day's potential LAI, stress factor, and stress
nocivity factor.
}
\examples{
# Example usage:
calcLeafAreaIndex(
  LAI = 2.5, LAIp = 3.0, LAIp_previousDay = 2.8, stress = 0.9,
  stress_nocivity = 1.2
)

}
