% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-calcLogisticTemperature.R
\name{calcLogisticTemperature}
\alias{calcLogisticTemperature}
\title{Calculate Logistic Temperature}
\usage{
calcLogisticTemperature(nTT, alpha1, beta)
}
\arguments{
\item{nTT}{\link{numeric} The input temperature value or thermal time.}

\item{alpha1}{\link{numeric} A parameter that influences the shape of the logistic
function.}

\item{beta}{\link{numeric} A parameter that scales the logistic function.}
}
\value{
\link{numeric} The calculated logistic temperature value.
}
\description{
This function computes a logistic temperature value based on the given
parameters.
}
\examples{
# Example usage:
calcLogisticTemperature(nTT = 2, alpha1 = 1.5, beta = 0.5)

}
