% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-calcNormalizedThermalTime.R
\name{calcNormalizedThermalTime}
\alias{calcNormalizedThermalTime}
\title{Calculate Normalized Thermal Time}
\usage{
calcNormalizedThermalTime(TT, TT_LAImax)
}
\arguments{
\item{TT}{\link{numeric} The thermal time value.}

\item{TT_LAImax}{\link{numeric} The maximum thermal time for leaf area index.}
}
\value{
\link{numeric} The normalized thermal time, which is the ratio of TT to
TT_LAImax.
}
\description{
This function calculates the normalized thermal time by dividing the thermal
time (TT) by the maximum thermal time for leaf area index (TT_LAImax).
}
\examples{
# Example usage:
calcNormalizedThermalTime(50, 100)
# [1] 0.5

}
