% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-calcLeafAreaIndex.R
\name{calcPotentialLeafAreaIndex}
\alias{calcPotentialLeafAreaIndex}
\title{Calculate Potential Leaf Area Index}
\usage{
calcPotentialLeafAreaIndex(TL, LAI0, LAImax)
}
\arguments{
\item{TL}{\link{numeric} Thermal time, a value between 0 and 1 representing the
proportion of the growing season completed.}

\item{LAI0}{\link{numeric} Initial leaf area index at the start of the growing
season.}

\item{LAImax}{\link{numeric} Maximum leaf area index achievable during the growing
season.}
}
\value{
\link{numeric} The potential leaf area index at the given thermal time.
}
\description{
This function calculates the potential leaf area index (LAI) based on the
thermal time (TL), initial leaf area index (LAI0), and maximum leaf area
index (LAImax).
}
\examples{
# Example usage:
calcPotentialLeafAreaIndex(TL = 0.5, LAI0 = 0.2, LAImax = 3.0)
# Returns: 1.6

}
