% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-calcThermalTime.R
\name{calcThermalTime}
\alias{calcThermalTime}
\title{Calculate Thermal Time}
\usage{
calcThermalTime(temp, temp_base)
}
\arguments{
\item{temp}{\link{numeric} vector of temperatures.}

\item{temp_base}{\link{numeric} value representing the base temperature.
Temperatures below this value will not contribute to the thermal time.}
}
\value{
\link{numeric} vector of the same length as \code{temp}, representing the
cumulative thermal time at each point.
}
\description{
This function calculates the cumulative thermal time based on a vector of
temperatures and a specified base temperature. Thermal time is calculated as
the cumulative sum of the positive differences between the temperature and
the base temperature.
}
\examples{
# Example usage:
temperatures <- c(10, 15, 20, 25)
base_temp <- 12
calcThermalTime(temperatures, base_temp)
# Output: 0 3 11 24

}
